+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/propertyComplianceViewModels/PropertyComplianceViewModel.kt
@@ -1,42 +1,83 @@
 package uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.propertyComplianceViewModels
 
+import uk.gov.communities.prsdb.webapp.annotations.PrsdbWebService
 import uk.gov.communities.prsdb.webapp.controllers.PropertyComplianceController
 import uk.gov.communities.prsdb.webapp.database.entity.PropertyCompliance
 import uk.gov.communities.prsdb.webapp.forms.steps.PropertyComplianceStepId
 import uk.gov.communities.prsdb.webapp.helpers.extensions.addRow
 import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.SummaryListRowViewModel
 
+// Simple data model with all public properties exposed as "val" constructor parameters
 class PropertyComplianceViewModel(
-    private val propertyCompliance: PropertyCompliance,
-    private val landlordView: Boolean = true,
+    val gasSafetySummaryList: List<SummaryListRowViewModel>,
+    val eicrSummaryList: List<SummaryListRowViewModel>,
+    val epcSummaryList: List<SummaryListRowViewModel>,
+    val landlordResponsibilitiesSummaryList: List<SummaryListRowViewModel>,
+    val landlordResponsibilitiesHintText: String,
+    val notificationMessages: List<PropertyComplianceNotificationMessage>,
 ) {
-    var notificationMessages: List<PropertyComplianceNotificationMessage> = getNotificationMessageKeys()
+    data class PropertyComplianceNotificationMessage(
+        val mainText: String,
+        val linkMessage: PropertyComplianceLinkMessage? = null,
+    )
 
-    val gasSafetySummaryList: List<SummaryListRowViewModel> =
-        GasSafetyViewModelBuilder.fromEntity(
-            propertyCompliance,
-            landlordView,
+    data class PropertyComplianceLinkMessage(
+        val linkUrl: String,
+        val linkText: String,
+        val afterLinkText: String,
+        val beforeLinkText: String? = null,
+        val isAfterLinkTextFullStop: Boolean = false,
+    )
+}
+
+@PrsdbWebService
+class PropertyComplianceViewModelFactory(
+    private val gasSafetyViewModelFactory: GasSafetyViewModelFactory,
+    private val eicrViewModelFactory: EicrViewModelFactory,
+) {
+    fun create(
+        propertyCompliance: PropertyCompliance,
+        landlordView: Boolean = true,
+    ): PropertyComplianceViewModel {
+        val gasSafetySummaryList: List<SummaryListRowViewModel> =
+            gasSafetyViewModelFactory.fromEntity(
+                propertyCompliance,
+                landlordView,
+            )
+
+        val eicrSummaryList: List<SummaryListRowViewModel> =
+            eicrViewModelFactory.fromEntity(propertyCompliance, landlordView)
+
+        val epcSummaryList: List<SummaryListRowViewModel> =
+            EpcViewModelBuilder.fromEntity(propertyCompliance, landlordView)
+
+        val landlordResponsibilitiesSummaryList: List<SummaryListRowViewModel> =
+            LandlordResponsibilitiesViewModelBuilder.fromEntity(
+                propertyCompliance,
+                landlordView,
+            )
+
+        val landlordResponsibilitiesHintText =
+            if (landlordView) {
+                "propertyDetails.complianceInformation.landlordResponsibilities.landlord.hintText"
+            } else {
+                "propertyDetails.complianceInformation.landlordResponsibilities.localAuthority.hintText"
+            }
+        return PropertyComplianceViewModel(
+            gasSafetySummaryList = gasSafetySummaryList,
+            eicrSummaryList = eicrSummaryList,
+            epcSummaryList = epcSummaryList,
+            landlordResponsibilitiesSummaryList = landlordResponsibilitiesSummaryList,
+            landlordResponsibilitiesHintText = landlordResponsibilitiesHintText,
+            notificationMessages = getNotificationMessageKeys(propertyCompliance, landlordView),
         )
+    }
 
-    val eicrSummaryList: List<SummaryListRowViewModel> = EicrViewModelBuilder.fromEntity(propertyCompliance, landlordView)
-
-    val epcSummaryList: List<SummaryListRowViewModel> = EpcViewModelBuilder.fromEntity(propertyCompliance, landlordView)
-
-    val landlordResponsibilitiesSummaryList: List<SummaryListRowViewModel> =
-        LandlordResponsibilitiesViewModelBuilder.fromEntity(
-            propertyCompliance,
-            landlordView,
-        )
-
-    val landlordResponsibilitiesHintText =
-        if (landlordView) {
-            "propertyDetails.complianceInformation.landlordResponsibilities.landlord.hintText"
-        } else {
-            "propertyDetails.complianceInformation.landlordResponsibilities.localAuthority.hintText"
-        }
-
-    private fun getNotificationMessageKeys(): List<PropertyComplianceNotificationMessage> =
-        mutableListOf<PropertyComplianceNotificationMessage>()
+    private fun getNotificationMessageKeys(
+        propertyCompliance: PropertyCompliance,
+        isLandlordView: Boolean,
+    ): List<PropertyComplianceViewModel.PropertyComplianceNotificationMessage> =
+        mutableListOf<PropertyComplianceViewModel.PropertyComplianceNotificationMessage>()
             .apply {
                 if (propertyCompliance.isGasSafetyCertExpired == true) {
                     addRow(
@@ -47,7 +88,7 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.gasCert.expired.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isGasSafetyCertMissing) {
@@ -59,7 +100,7 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.gasCert.missing.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isEicrExpired == true) {
@@ -71,7 +112,7 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.eicr.expired.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isEicrMissing) {
@@ -83,7 +124,7 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.eicr.missing.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isEpcExpired == true) {
@@ -95,7 +136,7 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.epc.expired.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isEpcRatingLow == true) {
@@ -109,7 +150,7 @@ class PropertyComplianceViewModel(
                         "propertyDetails.complianceInformation.notificationBanner.epc.lowRating.afterLinkText",
                         "propertyDetails.complianceInformation.notificationBanner.epc.lowRating.beforeLinkText",
                         isAfterLinkTextFullStop = true,
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
                 if (propertyCompliance.isEpcMissing) {
@@ -121,21 +162,8 @@ class PropertyComplianceViewModel(
                         ),
                         "propertyDetails.complianceInformation.notificationBanner.epc.missing.linkText",
                         "propertyDetails.complianceInformation.notificationBanner.asSoonAsPossible",
-                        withLinkMessage = landlordView,
+                        withLinkMessage = isLandlordView,
                     )
                 }
             }.toList()
-
-    data class PropertyComplianceNotificationMessage(
-        val mainText: String,
-        val linkMessage: PropertyComplianceLinkMessage? = null,
-    )
-
-    data class PropertyComplianceLinkMessage(
-        val linkUrl: String,
-        val linkText: String,
-        val afterLinkText: String,
-        val beforeLinkText: String? = null,
-        val isAfterLinkTextFullStop: Boolean = false,
-    )
 }
