+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/propertyComplianceViewModels/EicrViewModelBuilderTests.kt
@@ -23,7 +23,7 @@ class EicrViewModelBuilderTests {
         expectedRows: List<SummaryListRowViewModel>,
     ) {
         val eicrRows =
-            EicrViewModelBuilder.fromEntity(
+            EicrViewModelFactory().fromEntity(
                 propertyCompliance,
                 withActionLinks = withActionLinks,
             )
